requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
bcrypt==4.3.0
bidict==0.23.1
certifi==2025.8.3
cffi==2.0.0
click==8.3.0
cryptography==46.0.1
docutils==0.22.2
ecdsa==0.19.1
fastapi==0.117.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
jinja2==3.1.6
markdown2==2.5.4
markupsafe==3.0.2
multidict==6.6.4
nicegui==2.24.2
orjson==3.11.3
passlib==1.7.4
propcache==0.3.2
pscript==0.7.7
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-core==2.33.2
pygments==2.19.2
pyjwt==2.10.1
python-dotenv==1.1.1
python-engineio==4.12.2
python-jose==3.5.0
python-multipart==0.0.20
python-socketio==5.13.0
pyyaml==6.0.2
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.43
sqlmodel==0.0.25
starlette==0.48.0
typing-extensions==4.15.0
typing-inspection==0.4.1
uvicorn==0.37.0
uvloop==0.21.0
vbuild==0.8.2
watchfiles==1.1.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1
